volt meter example checkpoint
