+++ package.json
@@ -11,8 +11,10 @@
   ],
   "scripts": {
     "start": "bun run build && bun --hot run dev-server.ts",
+    "dev": "bun run start",
     "build": "bunx make-vfs ./symbols --content-format import-default --outfile ./generated/symbols-index.ts && tsup-node ./index.ts --format esm --dts --sourcemap && bun run scripts/generate-base-symbol-names.ts",
-    "generate": "bun run scripts/generate-symbols-from-asset-svgs.ts && bun run format",
+    "generate": "bun run scripts/generate-symbol-from-asset-svg.ts && bun run format",
+    "generate:all": "bun run scripts/generate-symbols-from-asset-svgs.ts && bun run format",
     "format:check": "bunx biome format",
     "format": "bunx biome format . --write",
     "vercel-build": "bun run ./scripts/build.ts"
@@ -20,8 +22,11 @@
   "devDependencies": {
     "@biomejs/biome": "^1.8.3",
     "@types/bun": "latest",
+    "@types/prompts": "^2.4.9",
     "@types/svg-path-parser": "^1.1.6",
+    "kleur": "^4.1.5",
     "make-vfs": "1.0.15",
+    "prompts": "^2.4.2",
     "svg-path-parser": "^1.1.0",
     "svgson": "^5.3.1",
     "transformation-matrix": "^2.16.1",
