+++ symbols/volt_meter_vert.ts
@@ -1,0 +1,4 @@
+import { rotateSymbol } from "drawing/rotateSymbol"
+import volt_meter_horz from "./volt_meter_horz"
+
+export default rotateSymbol(volt_meter_horz)
